UI touchups 2
